pyproject.toml
[project]
name = "clone-x"
version = "0.1.0"
description = ""
authors = [
    {name = "Mauro",email = "mauro_leao@hotmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "python-dotenv>=1.0",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "gunicorn>=21.0.0",
    "psycopg2-binary>=2.9.0",
    "dj-database-url>=2.0.0",
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-django = "^4.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
